feat: add bundled notifications skill and move send_notification from messaging
