+++ assistant/src/config/bundled-skills/messaging/TOOLS.json
@@ -155,87 +155,6 @@
       "execution_target": "host"
     },
     {
-      "name": "send_notification",
-      "description": "Send a notification through the unified notification router. The router decides channels/copy from preferences and urgency hints. Include a confidence score (0-1).",
-      "category": "messaging",
-      "risk": "medium",
-      "input_schema": {
-        "type": "object",
-        "properties": {
-          "message": {
-            "type": "string",
-            "description": "Notification message the user should receive"
-          },
-          "title": {
-            "type": "string",
-            "description": "Optional notification title"
-          },
-          "urgency": {
-            "type": "string",
-            "enum": [
-              "low",
-              "medium",
-              "high"
-            ],
-            "description": "Urgency hint (default: \"medium\")"
-          },
-          "requires_action": {
-            "type": "boolean",
-            "description": "Whether the notification expects user action (default: true)"
-          },
-          "is_async_background": {
-            "type": "boolean",
-            "description": "Whether the event is asynchronous/background work (default: false)"
-          },
-          "visible_in_source_now": {
-            "type": "boolean",
-            "description": "Set true when user is already viewing the source context (default: false)"
-          },
-          "deadline_at": {
-            "type": "number",
-            "description": "Optional deadline timestamp in epoch milliseconds"
-          },
-          "preferred_channels": {
-            "type": "array",
-            "items": {
-              "type": "string",
-              "enum": [
-                "vellum",
-                "telegram"
-              ]
-            },
-            "description": "Optional routing hints for preferred channels (not deterministic)"
-          },
-          "conversation_id": {
-            "type": "string",
-            "description": "Optional source conversation/session ID for notification context"
-          },
-          "source_event_name": {
-            "type": "string",
-            "description": "Optional event name for audit and dedupe grouping (default: \"user.send_notification\")"
-          },
-          "deep_link_metadata": {
-            "type": "object",
-            "description": "Optional metadata clients can use for deep linking"
-          },
-          "dedupe_key": {
-            "type": "string",
-            "description": "Optional dedupe key to suppress duplicate notifications"
-          },
-          "confidence": {
-            "type": "number",
-            "description": "Confidence score (0-1) for this action"
-          }
-        },
-        "required": [
-          "message",
-          "confidence"
-        ]
-      },
-      "executor": "tools/send-notification.ts",
-      "execution_target": "host"
-    },
-    {
       "name": "messaging_reply",
       "description": "Reply in a thread. Medium-risk action. Include a confidence score (0-1).",
       "category": "messaging",
