+++ assistant/src/config/bundled-skills/notifications/tools/shared.ts
@@ -1,0 +1,13 @@
+/**
+ * Shared utilities for notifications skill tools.
+ */
+
+import type { ToolExecutionResult } from '../../../../tools/types.js';
+
+export function ok(content: string): ToolExecutionResult {
+  return { content, isError: false };
+}
+
+export function err(message: string): ToolExecutionResult {
+  return { content: message, isError: true };
+}
