+++ assistant/src/config/bundled-tool-registry.ts
@@ -108,7 +108,8 @@ import * as messagingRead from './bundled-skills/messaging/tools/messaging-read.
 import * as messagingReply from './bundled-skills/messaging/tools/messaging-reply.js';
 import * as messagingSearch from './bundled-skills/messaging/tools/messaging-search.js';
 import * as messagingSend from './bundled-skills/messaging/tools/messaging-send.js';
-import * as sendNotification from './bundled-skills/messaging/tools/send-notification.js';
+// ── notifications ───────────────────────────────────────────────────────────
+import * as sendNotification from './bundled-skills/notifications/tools/send-notification.js';
 import * as sequenceAnalytics from './bundled-skills/messaging/tools/sequence-analytics.js';
 import * as sequenceCancel from './bundled-skills/messaging/tools/sequence-cancel.js';
 import * as sequenceCreate from './bundled-skills/messaging/tools/sequence-create.js';
@@ -250,7 +251,6 @@ export const bundledToolRegistry = new Map<string, SkillToolScript>([
   ['messaging:tools/messaging-read.ts', messagingRead],
   ['messaging:tools/messaging-search.ts', messagingSearch],
   ['messaging:tools/messaging-send.ts', messagingSend],
-  ['messaging:tools/send-notification.ts', sendNotification],
   ['messaging:tools/messaging-reply.ts', messagingReply],
   ['messaging:tools/messaging-mark-read.ts', messagingMarkRead],
   ['messaging:tools/slack-add-reaction.ts', slackAddReaction],
@@ -289,6 +289,9 @@ export const bundledToolRegistry = new Map<string, SkillToolScript>([
   ['messaging:tools/sequence-cancel.ts', sequenceCancel],
   ['messaging:tools/sequence-import.ts', sequenceImport],
   ['messaging:tools/sequence-analytics.ts', sequenceAnalytics],
+
+  // notifications
+  ['notifications:tools/send-notification.ts', sendNotification],
 
   // playbooks
   ['playbooks:tools/playbook-create.ts', playbookCreate],
